ml_service/requirements.txt
flask==2.3.3
gunicorn==21.2.0
scikit-learn==1.8.0
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.14.1
pillow==10.1.0
imagehash==4.3.1
opencv-python-headless==4.8.1.78
faiss-cpu==1.7.4
requests==2.31.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch